+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Automatically add default SSH key, generating if required
 - Stop it repeatedly updating Homebrew
 ### Changed
-- Handle casks AKA Mac Applications better
+- Handle casks (AKA Mac Applications) better
 ### Fixed
 - Issue with `gtc` command when running from a repository's subdirectory
 ### Removed