+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.6.3]
+### Fixed
+- Made gtc alias cope with the branch not being in the remote yet
+
 ## [1.6.2]
 ### Fixed
 - Fallback to using default Node.js when no .nvmrc instead of borking up